refactor(content.proto): change 'created_by' and 'updated_by' fields from int64 to string to store email ids instead of user ids
The 'created_by' and 'updated_by' fields in the ContentAttributes
message have been changed from int64 to string to store email ids
instead of user ids. This change allows for more flexibility in
identifying the users who created and updated the content. The
validation for 'created_by' and 'updated_by' fields in the
NewContentAttributes and ExistingContentAttributes messages has been
removed as it is not implemented yet.