chore(buf.gen.yaml): update go and go-grpc plugin versions, remove govalidators plugin
The go plugin version has been updated to v1.31.0 and the go-grpc plugin
version has been updated to v1.3.0. The govalidators plugin has been
removed from the list of plugins. These changes ensure that the
generated code is up to date and compatible with the latest versions of
the go and go-grpc plugins.