docs(README.md): improve readability and clarify development process
The changes in this commit improve the readability of the README.md file
by reformatting some lines to make them easier to read. Additionally,
the development process section is clarified to provide better
instructions for making changes to the project. Specifically, it is now
clear that changes should be made on a new branch, pushed to create a
pull request, and then merged into the master branch. The automated
process of running `buf lint` and generating updated Go code is also
mentioned. Finally, a note is added to remind contributors to run `buf
beta mod update` when updating the `buf.gen.yaml` file.