feat(feature_annotation.proto): add ignore option to updated_by field in NewFeatureAnnotation message
The updated_by field in the NewFeatureAnnotation message now has the `ignore` option set to `IGNORE_IF_UNPOPULATED`. This means that the validation rule for the email format will be skipped if the field is not populated. This change is necessary to allow the creation of new feature annotations without requiring an initial value for the `updated_by` field.