refactor(feature_annotation.proto): deprecate attributes field in FeatureAnnotationUpdate and add update_attributes field
The `attributes` field in `FeatureAnnotationUpdate` is deprecated in
favor of `update_attributes`. This change allows for partial updates to
feature annotations by making all fields within `update_attributes`
optional. The `attributes` field is maintained for backward
compatibility but will be removed in a future version. When both fields
are provided, `update_attributes` will take precedence.